chore: utils module containing write_sep (#1053)
* chore: utils module containing write_sep

* add doc
