+++ crates/iota-sdk-types/src/transaction/mod.rs
@@ -6,6 +6,7 @@ use super::{
     Address, CheckpointTimestamp, Digest, EpochId, Event, GenesisObject, Identifier, ObjectId,
     ObjectReference, ProtocolVersion, RandomnessRound, TypeTag, UserSignature, Version,
 };
+use crate::utils::write_sep;
 
 #[cfg(feature = "serde")]
 #[cfg_attr(doc_cfg, doc(cfg(feature = "serde")))]
@@ -1022,32 +1023,6 @@ impl Command {
             ticket,
         })
     }
-}
-
-/// Writes `items` to `f`, separated by `sep` and optionally wrapped in
-/// `delimiters` (e.g. `("[", "]")`). If the iterator is empty, nothing is
-/// written — not even the delimiters.
-pub fn write_sep<T: core::fmt::Display>(
-    f: &mut core::fmt::Formatter<'_>,
-    items: impl IntoIterator<Item = T>,
-    delimiters: Option<(&str, &str)>,
-    sep: &str,
-) -> std::fmt::Result {
-    let mut xs = items.into_iter();
-    let Some(x) = xs.next() else {
-        return Ok(());
-    };
-    if let Some((l, _)) = delimiters {
-        write!(f, "{l}")?;
-    }
-    write!(f, "{x}")?;
-    for x in xs {
-        write!(f, "{sep}{x}")?;
-    }
-    if let Some((_, r)) = delimiters {
-        write!(f, "{r}")?;
-    }
-    Ok(())
 }
 
 impl core::fmt::Display for MoveCall {
